pyproject.toml
[project]
name = "justsayit"
version = "0.5.0"
description = "Local Parakeet-powered voice dictation with a Wayland layer-shell overlay"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "horo" }]
dependencies = [
    "sherpa-onnx>=1.10",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "platformdirs>=4.2",
    "PyGObject>=3.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
whisper = [
    "faster-whisper>=1.0",
]
# NOTE: llama-cpp-python must be compiled with Vulkan support for GPU use.
# install.sh --postprocess handles this automatically.
# For manual install: CMAKE_ARGS="-DGGML_VULKAN=1" pip install llama-cpp-python
llm = [
    "llama-cpp-python>=0.3",
]

[project.scripts]
justsayit = "justsayit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/justsayit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.uv]
# PyGObject builds against system headers; rely on system-site-packages venv
# for gi.repository bindings (gtk4, gtk4-layer-shell) already installed via pacman.
